requirements.txt
streamlit
pandas
requests
plotly
streamlit-lottie
psycopg2-binary
python-dotenv
scikit-learn
numpy
bcrypt
openpyxl
openai
matplotlib
tabulate
beautifulsoup4
google-generativeai 
matplotlib